+++ src/app/reliquia/AnilloVilya.java
@@ -2,7 +2,7 @@ package app.reliquia;
 
 public class AnilloVilya extends AnilloElfo{
 
-    public AnilloVilya(String nombre, int factorDeAtaque, int factorDeDefensa, int energiaMagica) {
+    public AnilloVilya(String nombre, double factorDeAtaque, double factorDeDefensa, int energiaMagica) {
         super(nombre, factorDeAtaque, factorDeDefensa, energiaMagica);
     }
 }
